product-url-crawler/url-extractor/requirements.txt
functions-framework==3.*
requests==2.*
google-cloud-storage==2.*
google-cloud-tasks==2.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*